Revert "Fixing all models to match up with our current data model"

+++ src/main/java/team/awesome/pet/service/PetRescueOrganizationService.java
@@ -1,6 +1,5 @@
 package team.awesome.pet.service;
 
-import java.math.BigInteger;
 import java.util.Collection;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -101,7 +100,7 @@ public class PetRescueOrganizationService implements PetRescueOrganizationDao {
   }
 
   @Override
-  public void removeShelterById(BigInteger id) {
+  public void removeShelterById(int id) {
     this.shelterDao.removeShelterById(id);
   }
